+++ CHANGELOG.md
@@ -19,13 +19,22 @@ and this project adheres to
 
 - Add a mix task to merge projects
   [#3615](https://github.com/OpenFn/lightning/issues/3615)
+- Show validation error message when creating sandbox with duplicate name
+  [#3776](https://github.com/OpenFn/lightning/issues/3776)
 - Divergence warning when merging sandboxes - displays alert if target branch
   was modified after sandbox creation to prevent data loss
   [#3747](https://github.com/OpenFn/lightning/issues/3747)
+- Sandbox indicator banners in workflow editor (inspector) to help indicate when
+  working in a sandbox environment
+  [#3413](https://github.com/OpenFn/lightning/issues/3413)
 
 ### Changed
 
 ### Fixed
+
+- Fixed sandbox merge failing with StaleEntryError when parent workflow was
+  modified after sandbox creation
+  [#3765](https://github.com/OpenFn/lightning/issues/3765)
 
 ## [2.14.13-pre1] 2025-10-24